Change the type of result returned by XPA.get and XPA.set
`XPA.get` and `XPA.set` now return an instance of `Reply`.  Methods are
provided to deal with such instances and to retrieve the contents (name of
server, message or data) of a given reply in various forms.  Instances of
`Reply` are mutable object and have a finalizer which takes care of correctly
freeing memory.  By default, when retrieving the data associated with a given
reply, the ownership of the data is transfered to Julia thus avoiding
unecessary copies.

Old `XPA.get_XXXX` methods have been suppressed as the same result can be
obtained by spcifying the type of result in calls to `XPA.get` or to the new
`XPA.get_data`.